+++ raven-data-entity/src/main/java/raven/data/entity/AutoIncr.java
@@ -1,0 +1,8 @@
+package raven.data.entity;
+
+/**
+ *
+ * */
+public interface AutoIncr<TKey> extends Entity<TKey> {
+}
+
