+++ raven-mongodb-repository/pom.xml
@@ -2,34 +2,39 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>raven</artifactId>
+        <groupId>org.raven</groupId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.raven</groupId>
-    <artifactId>raven</artifactId>
-    <packaging>pom</packaging>
-    <version>1.0-SNAPSHOT</version>
-    <modules>
-        <module>raven-data-entity</module>
-        <module>raven-mongodb-repository</module>
-    </modules>
+    <groupId>org.raven.mongodb.repository</groupId>
+    <artifactId>raven-mongodb-repository</artifactId>
+
 
     <dependencies>
+
         <dependency>
-            <groupId>org.mongodb</groupId>
-            <artifactId>mongodb-driver</artifactId>
-            <version>3.6.0</version>
+            <groupId>org.raven.data.entity</groupId>
+            <artifactId>raven-data-entity</artifactId>
+            <version>${project.version}</version>
         </dependency>
+
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
             <version>4.12</version>
             <scope>test</scope>
         </dependency>
+
         <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>2.5.3</version>
+            <groupId>org.mongodb</groupId>
+            <artifactId>mongodb-driver</artifactId>
+            <version>3.6.0</version>
         </dependency>
+
+
     </dependencies>
 
 </project>
