+++ raven-mongodb-repository/src/main/java/raven/mongodb/repository/MongoSequence.java
