+++ raven-mongodb-repository/src/main/java/raven/mongodb/repository/exceptions/EmptyArgumentException.java
