+++ raven-mongodb-repository/src/main/java/raven/mongodb/repository/exceptions/FailedException.java
