+++ raven-mongodb-repository/src/test/java/raven/mongodb/repository/MallRepositoryImpl.java
@@ -1,0 +1,11 @@
+package raven.mongodb.repository;
+
+import com.mongodb.WriteConcern;
+
+public class MallRepositoryImpl extends MongoRepositoryImpl<Mall, String> {
+    public MallRepositoryImpl() {
+        super("mongodb://127.0.0.1:27017/"
+                , "RepositoryTest2", null, WriteConcern.ACKNOWLEDGED, null, null);
+
+    }
+}
