+++ raven-mongodb-repository/src/test/java/raven/mongodb/repository/User2RepositoryImpl.java
@@ -1,0 +1,11 @@
+package raven.mongodb.repository;
+
+import com.mongodb.WriteConcern;
+
+public class User2RepositoryImpl extends MongoRepositoryImpl<User2, Long> {
+    public User2RepositoryImpl() {
+        super("mongodb://127.0.0.1:27017/"
+                , "RepositoryTest2", "User", WriteConcern.ACKNOWLEDGED, null, null);
+
+    }
+}
