+++ raven-mongodb-repository/src/test/java/raven/mongodb/repository/UserRepositoryImpl.java
@@ -1,0 +1,11 @@
+package raven.mongodb.repository;
+
+import com.mongodb.WriteConcern;
+
+public class UserRepositoryImpl extends MongoRepositoryImpl<User, Long> {
+    public UserRepositoryImpl() {
+        super("mongodb://127.0.0.1:27017/"
+                , "RepositoryTest2", null, WriteConcern.ACKNOWLEDGED, null, null);
+
+    }
+}
