--- src/main/java/raven/data/entity/AutoIncr.java
@@ -1,7 +1,0 @@
-package raven.data.entity;
-
-/**
- *
- * */
-public interface AutoIncr<TKey> extends Entity<TKey> {
-}
